FlaNET/MLServer/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
astroid==2.5.1
astunparse==1.6.3
beautifulsoup4==4.9.3
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cmdstanpy==0.9.5
colorama==0.4.4
convertdate==2.3.1
cryptography==3.4.6
cycler==0.10.0
Cython==0.29.22
ephem==3.7.7.1
fastapi==0.63.0
fbprophet==0.7.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
hijri-converter==2.1.1
holidays==0.10.5.2
idna==2.10
isort==5.7.0
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
korean-lunar-calendar==0.2.1
lazy-object-proxy==1.5.2
LunarCalendar==0.0.9
Markdown==3.3.4
matplotlib==3.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.3
pathspec==0.8.1
Pillow==8.1.2
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
pylint==2.7.2
PyMeeus==0.5.9
PyMySQL==1.0.2
pyparsing==2.4.7
pystan==2.19.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.1
setuptools-git==1.2
six==1.15.0
sklearn==0.0
soupsieve==2.2
SQLAlchemy==1.3.23
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.59.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
